REMOVE THE EXPECTATION OF A WP-CLI ENV FILE PASSED TO THE WP-CLI SERVICE
Since we defaulted the UID ARG in the Dockerfile, pasing an environment
file through to the wp-cli service became optional. If it isn't provided
then the wp-cli service will use the UID 1000 by default to write files
back to the client. This can still be overridden by project's to pass
a different value, but that would be in the project's Docker Compose
file rather than the one in this tool.